fix(#591): address review round 2 findings
- F1 fixed: extracted native cell traversal into shared/ghosttyCellTraversal.ts and imported it from both Electron main and the renderer.
- F2 fixed: replaced trim-based styled blank fallback consumption with adjacent-cell-aware traversal and added sparse whitespace regressions.

Reviewers: github-claude, github-codex-connector

Verification:
- npx vitest --run --passWithNoTests electron/ghostty-render-state-main.test.ts src/features/terminal/components/TerminalPane/ghosttyVtRenderSnapshot.test.ts shared/ghosttyCellTraversal.test.ts
- npx eslint electron/ghostty-render-state-main.ts src/features/terminal/components/TerminalPane/ghosttyVtRenderSnapshot.ts shared/ghosttyCellTraversal.ts shared/ghosttyCellTraversal.test.ts
- npx tsc -p electron/tsconfig.json --noEmit
- npx prettier --check tsconfig.json electron/ghostty-render-state-main.ts src/features/terminal/components/TerminalPane/ghosttyVtRenderSnapshot.ts shared/ghosttyCellTraversal.ts shared/ghosttyCellTraversal.test.ts
- codex verify cycle 2: pass

GitHub-Review-Processed-Claude: 4761081917
GitHub-Review-Superseded-Claude: 4761026660
GitHub-Review-Processed-Codex-Reviews: 4539263301
GitHub-Review-Processed-Codex-Inline: 3447979986
GitHub-Review-Processed-Human-Issue:
GitHub-Review-Processed-Human-Inline:
Closes-Codex-Threads: PRRT_kwDOR06LW86LEYSI
Pattern-Files-Touched:
  docs/reviews/patterns/terminal-render-state-driver-contract.md
  docs/reviews/CLAUDE.md
Pattern-Append-Decisions:
  - F1 (duplicated native cell traversal) -> patterns/terminal-render-state-driver-contract.md (existing, theme: native render-state traversal contract)
  - F2 (sparse styled blank whitespace fallback) -> patterns/terminal-render-state-driver-contract.md (existing, theme: native render-state fallback invariants)
